+++ src/app/components/installnav/installnav.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-installnav',
+  templateUrl: './installnav.component.html',
+  styleUrls: ['./installnav.component.css']
+})
+export class InstallnavComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}